+++ demo/src/App.js
@@ -11,6 +11,7 @@ const MIN_ZOOM = -8;
 const DEFAULT_VIEW_STATE = { zoom: -5.5, target: [30000, 10000, 0] };
 const COLORMAP = 'viridis';
 const COLORMAP_SLIDER_CHECKBOX_COLOR = [220, 220, 220];
+const MARGIN = 25;
 
 const initSourceName = 'zarr';
 const colorValues = [
@@ -34,6 +35,7 @@ function App() {
   const [colormapOn, setColormap] = useState('');
   const [zStack, setZStack] = useState(0);
   const [mzIndex, setMZIndex] = useState(0);
+  const [overviewOn, setOverview] = useState(false);
 
   useEffect(() => {
     const handleResize = () => {
@@ -112,6 +114,12 @@ function App() {
       return nextLoader;
     });
     setMZIndex(value);
+  };
+
+  const toggleOverview = () => {
+    setOverview(prevOverviewOn => {
+      return !prevOverviewOn;
+    });
   };
 
   const sourceButtons = Object.keys(sources).map(name => {
@@ -261,6 +269,14 @@ function App() {
           )}
           initialViewState={initialViewState}
           colormap={colormapOn}
+          overview={
+            overviewOn
+              ? {
+                  margin: MARGIN,
+                  scale: 0.2
+                }
+              : null
+          }
         />
       ) : null}
       <div className="slider-container">
@@ -291,6 +307,19 @@ function App() {
             {colormapOn ? 'Colors' : COLORMAP}
           </Button>
         </div>
+        {loader && loader.isPyramid ? (
+          <div style={{ marginTop: '15px', marginBottom: '15px' }}>
+            <Button
+              variant="contained"
+              onClick={() => toggleOverview()}
+              key="overview"
+            >
+              {overviewOn ? 'Remove Overview' : 'Show Overview'}
+            </Button>
+          </div>
+        ) : (
+          []
+        )}
         {sliders}
         {stackControllers}
       </div>